refactor: split FileOperationsService into FileDialogHandler and RecipeCommandHandler
