+++ src/App/ViewManager.cs
@@ -1,10 +1,12 @@
 using System.Diagnostics;
 using System.Diagnostics.CodeAnalysis;
 using System.Text;
+using DataMorph.App.Views;
 using DataMorph.Engine.IO;
 using DataMorph.Engine.IO.JsonLines;
 using DataMorph.Engine.Models;
 using DataMorph.Engine.Models.Actions;
+using DataMorph.Engine.Types;
 using Terminal.Gui.ViewBase;
 using Terminal.Gui.Views;
 
@@ -19,18 +21,79 @@ internal sealed class ViewManager : IDisposable
 {
     private readonly Window _container;
     private readonly AppState _state;
-    private readonly Func<Task> _onToggle;
+    private readonly ModeController _modeController;
     private View? _currentView;
     private bool _disposed;
 
-    internal ViewManager(Window container, AppState state, Func<Task> onToggle)
+    internal ViewManager(Window container, AppState state, ModeController modeController)
     {
         ArgumentNullException.ThrowIfNull(container);
         ArgumentNullException.ThrowIfNull(state);
-        ArgumentNullException.ThrowIfNull(onToggle);
+        ArgumentNullException.ThrowIfNull(modeController);
         _container = container;
         _state = state;
-        _onToggle = onToggle;
+        _modeController = modeController;
+    }
+
+    /// <summary>
+    /// Refreshes the status bar hints based on the current application state.
+    /// </summary>
+    internal void RefreshStatusBarHints()
+    {
+        var statusBar = GetCurrentStatusBar();
+        if (statusBar is null)
+        {
+            return;
+        }
+
+        List<string> hints = ["o:Open", "s:Save", "q:Quit"];
+
+        if (!string.IsNullOrWhiteSpace(_state.CurrentFilePath))
+        {
+            var format = FormatDetector.Detect(_state.CurrentFilePath);
+            if (format.IsSuccess && format.Value == DataFormat.JsonLines)
+            {
+                hints.Add("t:Tree/Table");
+            }
+
+            if (GetCurrentView() is MorphTableView)
+            {
+                hints.Add("x:Menu");
+            }
+        }
+
+        hints.Add("?:Help");
+        statusBar.Text = string.Join("  ", hints);
+    }
+
+    /// <summary>
+    /// Toggles between JSON Lines Tree and Table view modes.
+    /// </summary>
+    /// <returns>A task representing the asynchronous operation.</returns>
+    internal async Task ToggleJsonLinesModeAsync()
+    {
+        var result = await _modeController.ToggleJsonLinesModeAsync();
+
+        if (result.IsFailure)
+        {
+            ShowError(result.Error);
+            RefreshStatusBarHints();
+            return;
+        }
+
+        if (_state.CurrentMode == ViewMode.JsonLinesTree && _state.RowIndexer is not null)
+        {
+            SwitchToJsonLinesTree(_state.RowIndexer);
+            return;
+        }
+
+        if (
+            _state.CurrentMode == ViewMode.JsonLinesTable
+            && _state.RowIndexer is not null
+            && _state.Schema is not null)
+        {
+            SwitchToJsonLinesTableView(_state.RowIndexer, _state.Schema);
+        }
     }
 
     /// <summary>
@@ -40,6 +103,7 @@ internal sealed class ViewManager : IDisposable
     {
         ObjectDisposedException.ThrowIf(_disposed, this);
         SwapView(Views.FileSelectionView.Create());
+        RefreshStatusBarHints();
     }
 
     /// <summary>
@@ -92,6 +156,7 @@ internal sealed class ViewManager : IDisposable
             GetRawColumnName = getRawColumnName,
         };
         SwapView(view);
+        RefreshStatusBarHints();
 
         if (source is Views.LazyTransformer { FilterRowIndexer: { } filterIndexer })
         {
@@ -113,7 +178,7 @@ internal sealed class ViewManager : IDisposable
         ObjectDisposedException.ThrowIf(_disposed, this);
         ArgumentNullException.ThrowIfNull(indexer);
 
-        var view = new Views.JsonLinesTreeView(indexer, () => _ = _onToggle())
+        var view = new Views.JsonLinesTreeView(indexer, () => _ = ToggleJsonLinesModeAsync())
         {
             X = 0,
             Y = 0,
@@ -121,6 +186,7 @@ internal sealed class ViewManager : IDisposable
             Height = Dim.Fill(),
         };
         SwapView(view);
+        RefreshStatusBarHints();
     }
 
     /// <summary>
@@ -178,6 +244,7 @@ internal sealed class ViewManager : IDisposable
             GetRawColumnName = getRawColumnName,
         };
         SwapView(view);
+        RefreshStatusBarHints();
 
         if (tableSource is Views.LazyTransformer { FilterRowIndexer: { } filterIndexer })
         {
