+++ tests/DataMorph.Tests/App/FileDialogHandlerTests.cs
@@ -1,0 +1,40 @@
+using AwesomeAssertions;
+using DataMorph.App;
+using Terminal.Gui.App;
+using Terminal.Gui.Drivers;
+using Terminal.Gui.Views;
+
+namespace DataMorph.Tests.App;
+
+public sealed class FileDialogHandlerTests
+{
+    private static IApplication CreateTestApp()
+    {
+        var app = Application.Create();
+        app.Init(DriverRegistry.Names.ANSI);
+        return app;
+    }
+
+    [Fact]
+    public void Constructor_WithValidDependencies_DoesNotThrow()
+    {
+        // Arrange
+        using var app = CreateTestApp();
+        using var state = new AppState();
+        using var window = new Window();
+        var modeController = new ModeController(state);
+        using var viewManager = new ViewManager(window, state, modeController);
+
+        // Act
+        Action act = () =>
+        {
+            var handler = new FileDialogHandler(app, state, viewManager, _ => { });
+        };
+
+        // Assert
+        act.Should().NotThrow();
+
+        // TODO: Testing the dialog-cancellation path requires injecting an IFileDialog abstraction.
+        // That refactoring is out of scope here; a future task should address this.
+    }
+}
