--- tests/DataMorph.Tests/App/FileOperationsServiceTests.cs
@@ -1,168 +1,0 @@
-using DataMorph.App;
-using DataMorph.App.Views;
-using Terminal.Gui.App;
-using Terminal.Gui.Drivers;
-using Terminal.Gui.Views;
-
-namespace DataMorph.Tests.App;
-
-public sealed class FileOperationsServiceTests
-{
-    private static IApplication CreateTestApp()
-    {
-        var app = Application.Create();
-        app.Init(DriverRegistry.Names.ANSI);
-        return app;
-    }
-
-    [Fact]
-    public void UpdateStatusBarHints_WithNoFilePath_UsesDefaultHints()
-    {
-        // Arrange
-        using var app = CreateTestApp();
-        using var state = new AppState { CurrentFilePath = string.Empty };
-        using var window = new Window();
-        using var viewManager = new ViewManager(window, state, () => Task.CompletedTask);
-        var modeController = new ModeController(state);
-        var service = new FileOperationsService(app, state, viewManager, modeController);
-
-        // Act
-        service.UpdateStatusBarHints(null);
-
-        // Assert
-        // Default hints should be set: "o:Open", "s:Save", "q:Quit", "?:Help"
-        // We can't directly verify statusBar.Text without exposing it,
-        // but we can verify the method completes without errors
-    }
-
-    [Fact]
-    public void UpdateStatusBarHints_WithCsvFilePath_UsesDefaultHints()
-    {
-        // Arrange
-        using var app = CreateTestApp();
-        using var state = new AppState { CurrentFilePath = "test.csv" };
-        using var window = new Window();
-        using var viewManager = new ViewManager(window, state, () => Task.CompletedTask);
-        var modeController = new ModeController(state);
-        var service = new FileOperationsService(app, state, viewManager, modeController);
-
-        // Act
-        service.UpdateStatusBarHints(null);
-
-        // Assert
-        // CSV files should only show default hints: "o:Open", "s:Save", "q:Quit", "?:Help"
-    }
-
-    [Fact]
-    public void UpdateStatusBarHints_WithJsonLinesPath_IncludesToggleHint()
-    {
-        // Arrange
-        using var app = CreateTestApp();
-        using var state = new AppState { CurrentFilePath = "test.jsonl" };
-        using var window = new Window();
-        using var viewManager = new ViewManager(window, state, () => Task.CompletedTask);
-        var modeController = new ModeController(state);
-        var service = new FileOperationsService(app, state, viewManager, modeController);
-
-        // Act
-        service.UpdateStatusBarHints(null);
-
-        // Assert
-        // JSON Lines files should include "t:Tree/Table" hint
-    }
-
-    [Fact]
-    public void UpdateStatusBarHints_WithMorphTableView_IncludesMenuHint()
-    {
-        // Arrange
-        using var app = CreateTestApp();
-        using var state = new AppState { CurrentFilePath = "test.jsonl" };
-        using var window = new Window();
-        using var viewManager = new ViewManager(window, state, () => Task.CompletedTask);
-
-        using var testTableView = new TestTableView
-        {
-            Table = new TestTableSource(),
-            GetRawColumnName = _ => "test",
-            OnMorphAction = _ => { }
-        };
-        window.Add(testTableView);
-
-        var modeController = new ModeController(state);
-        var service = new FileOperationsService(app, state, viewManager, modeController);
-
-        // Act
-        service.UpdateStatusBarHints(null);
-
-        // Assert
-        // Should include "x:Menu" hint when MorphTableView is present
-    }
-
-    [Fact]
-    public async Task HandleSaveRecipeAsync_WithNonTableMode_DoesNothing()
-    {
-        // Arrange
-        using var app = CreateTestApp();
-        using var state = new AppState { CurrentMode = ViewMode.FileSelection };
-        using var window = new Window();
-        using var viewManager = new ViewManager(window, state, () => Task.CompletedTask);
-        var modeController = new ModeController(state);
-        var service = new FileOperationsService(app, state, viewManager, modeController);
-
-        // Act
-        await service.HandleSaveRecipeAsync();
-
-        // Assert
-        // Should return early without showing dialog
-        // (No exception should be thrown)
-    }
-
-    [Fact]
-    public async Task HandleLoadRecipeAsync_WithNoFilePath_DoesNothing()
-    {
-        // Arrange
-        using var app = CreateTestApp();
-        using var state = new AppState { CurrentFilePath = string.Empty };
-        using var window = new Window();
-        using var viewManager = new ViewManager(window, state, () => Task.CompletedTask);
-        var modeController = new ModeController(state);
-        var service = new FileOperationsService(app, state, viewManager, modeController);
-
-        // Act
-        await service.HandleLoadRecipeAsync();
-
-        // Assert
-        // Should return early without showing dialog
-        // (No exception should be thrown)
-    }
-
-    /// <summary>
-    /// Testable concrete implementation of MorphTableView.
-    /// </summary>
-    private sealed class TestTableView : MorphTableView
-    {
-        public new ITableSource? Table { get; set; }
-    }
-
-    /// <summary>
-    /// Simple TableSource implementation for testing.
-    /// </summary>
-    private sealed class TestTableSource : ITableSource
-    {
-        public int Rows => 10;
-        public int Columns => 3;
-        public string[] ColumnNames => ["Col1", "Col2", "Col3"];
-
-        public object this[int row, int col]
-        {
-            get => $"R{row}C{col}";
-            set { }
-        }
-
-        public static void AddColumn(string name) { }
-        public static void AddRow() { }
-        public static void RemoveColumn(int index) { }
-        public static void RemoveRow(int index) { }
-        public static void Clear() { }
-    }
-}
